[refactor]callメソッドの実装
