+++ app/controllers/songs/songs_controller.rb
@@ -1,13 +1,13 @@
 class Songs::SongsController < ApplicationController
   def show
-    @song_id = params[:id]
-    @share_id = params[:share_id]
-
-    service = ShareSongService.new(song_id: @song_id, share_id: @share_id)
-
-    song = ShareSongService.fetch_song_and_artist_name(@song_id)
-    render json: song
-
-    service.increment_shared_count
+    song_data = FetchSongService.call(
+      song_id: params[:id],
+      share_id: params[:share_id]
+    )
+    if song_data
+      render json: song_data, status: :ok
+    else
+      render json: { error: 'Song not found' }, status: :not_found
+    end
   end
 end
